backend/pyproject.toml
[project]
name = "ai-agent-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "asyncpg>=0.30.0",
    "geopy>=2.4.1",
    "google-api-python-client>=2.176.0",
    "google-auth>=2.40.3",
    "google-auth-httplib2>=0.2.0",
    "google-auth-oauthlib>=1.2.2",
    "ipython>=9.4.0",
    "langchain-google-genai>=2.1.7",
    "langchain-openai>=0.3.28",
    "langgraph>=0.5.2",
    "langgraph-supervisor>=0.0.27",
    "mem0ai>=0.1.114",
    "psycopg2-binary>=2.9.10",
    "python-dotenv>=1.1.1",
    "requests>=2.32.4",
    "sqlalchemy>=2.0.41",
]
